requirements.txt
torch~=1.13.1
torchvision~=0.14.1
Pillow~=9.5.0
timm~=0.4.5
numpy~=1.21.6
Django~=3.2.23
requests~=2.28.2
opencv-python~=4.5.5.64
matplotlib~=3.5.3
pandas~=1.3.5